fix: squad tour modal
